[Refactor] Refactor Some TonicsCloud Controllers into Services, and a couple of Optimize import
