+++ bin/restart_service.json
@@ -1,1 +1,1 @@
-{"info":"The timestamp would be changed whenever an app\/module is updated, deleted, installed, this way, you can restart any services running in bin\/console","timestamp":1715801617}
+{"info":"The timestamp would be changed whenever an app\/module is updated, deleted, installed, this way, you can restart any services running in bin\/console","timestamp":1715880806}
