+++ src/Apps/TonicsCloud/EventHandlers/Fields/CloudCredit.php
@@ -19,13 +19,8 @@
 namespace App\Apps\TonicsCloud\EventHandlers\Fields;
 
 use App\Apps\TonicsCloud\Controllers\BillingController;
-use App\Apps\TonicsCloud\Controllers\InstanceController;
-use App\Apps\TonicsCloud\Controllers\TonicsCloudSettingsController;
-use App\Apps\TonicsCloud\Interfaces\CloudServerInterface;
-use App\Apps\TonicsCloud\TonicsCloudActivator;
 use App\Modules\Field\Events\OnFieldMetaBox;
 use Devsrealm\TonicsEventSystem\Interfaces\HandlerInterface;
-use Devsrealm\TonicsQueryBuilder\TonicsQuery;
 
 class CloudCredit implements HandlerInterface
 {
