+++ src/Apps/TonicsCloud/EventHandlers/Fields/CloudInstanceInfo.php
@@ -20,11 +20,9 @@ namespace App\Apps\TonicsCloud\EventHandlers\Fields;
 
 use App\Apps\TonicsCloud\Controllers\InstanceController;
 use App\Apps\TonicsCloud\Controllers\TonicsCloudSettingsController;
-use App\Apps\TonicsCloud\Interfaces\CloudServerInterface;
 use App\Apps\TonicsCloud\TonicsCloudActivator;
 use App\Modules\Field\Events\OnFieldMetaBox;
 use Devsrealm\TonicsEventSystem\Interfaces\HandlerInterface;
-use Devsrealm\TonicsQueryBuilder\TonicsQuery;
 use Throwable;
 
 class CloudInstanceInfo implements HandlerInterface
