+++ src/Apps/TonicsCloud/EventHandlers/Fields/CloudInstances.php
@@ -19,13 +19,8 @@
 namespace App\Apps\TonicsCloud\EventHandlers\Fields;
 
 use App\Apps\TonicsCloud\Controllers\InstanceController;
-use App\Apps\TonicsCloud\Controllers\TonicsCloudSettingsController;
-use App\Apps\TonicsCloud\Interfaces\CloudServerInterface;
-use App\Apps\TonicsCloud\TonicsCloudActivator;
-use App\Modules\Core\Library\Authentication\Session;
 use App\Modules\Field\Events\OnFieldMetaBox;
 use Devsrealm\TonicsEventSystem\Interfaces\HandlerInterface;
-use Devsrealm\TonicsQueryBuilder\TonicsQuery;
 
 class CloudInstances implements HandlerInterface
 {
