+++ src/Apps/TonicsCloud/Jobs/Instance/CloudJobQueueInstanceHasDeleted.php
@@ -18,8 +18,6 @@
 
 namespace App\Apps\TonicsCloud\Jobs\Instance;
 
-use App\Apps\TonicsCloud\Controllers\TonicsCloudSettingsController;
-use App\Apps\TonicsCloud\Interfaces\CloudServerInterface;
 use App\Apps\TonicsCloud\Jobs\Instance\Traits\TonicsJobQueueInstanceTrait;
 use App\Apps\TonicsCloud\TonicsCloudActivator;
 use App\Modules\Core\Library\JobSystem\AbstractJobInterface;
