+++ src/Apps/TonicsCloud/Jobs/Instance/CloudJobQueueStartInstance.php
@@ -18,10 +18,8 @@
 
 namespace App\Apps\TonicsCloud\Jobs\Instance;
 
-use App\Apps\TonicsCloud\Controllers\TonicsCloudSettingsController;
 use App\Apps\TonicsCloud\Interfaces\CloudServerInterface;
 use App\Apps\TonicsCloud\Jobs\Instance\Traits\TonicsJobQueueInstanceTrait;
-use App\Apps\TonicsCloud\TonicsCloudActivator;
 use App\Modules\Core\Library\JobSystem\AbstractJobInterface;
 use App\Modules\Core\Library\JobSystem\JobHandlerInterface;
