+++ src/Apps/TonicsCloud/RequestInterceptor/TonicsCloudDomainAccess.php
@@ -18,7 +18,6 @@
 
 namespace App\Apps\TonicsCloud\RequestInterceptor;
 
-use App\Apps\TonicsCloud\Controllers\ContainerController;
 use App\Apps\TonicsCloud\Controllers\DomainController;
 use App\Modules\Core\Library\SimpleState;
 use Devsrealm\TonicsRouterSystem\Events\OnRequestProcess;
