+++ src/Apps/TonicsCloud/Views/Container/edit.html
@@ -10,6 +10,7 @@
 
 [[hook_into('Core::before_form_section')
     <input type="hidden" name="slug_id" value="[[v('ContainerData.slug_id')]]">
+    <input type="hidden" name="container_id" value="[[v('ContainerData.container_id')]]">
 ]]
 
 [[hook_into('Core::before_footer')
