+++ src/Modules/Core/Controllers/OEmbedController.php
@@ -25,6 +25,7 @@ class OEmbedController
 {
     /**
      * @throws \Exception
+     * @throws \Throwable
      */
     public function OEmbed()
     {
